+++ include/material.hpp
@@ -6,20 +6,20 @@
 
 class Material: public std::enable_shared_from_this<Material>{
 protected:
-    double m_reflectiveness;
+    float m_reflectiveness;
 public:
-	Material(double reflectiveness):m_reflectiveness(reflectiveness){};
+	Material(float reflectiveness):m_reflectiveness(reflectiveness){};
     virtual PtrColor sample(PtrRay ray, PtrVector position, PtrVector normal) = 0;
-    inline double getReflectiveness(){ return m_reflectiveness; };
+    inline float getReflectiveness(){ return m_reflectiveness; };
 public:
     static PtrVector LightDir;
     static PtrColor LightColor;
 };
 
 class CheckerMaterial: public Material{
-    double m_scale;
+    float m_scale;
 public:
-    CheckerMaterial(double scale, double reflectiveness=0);
+    CheckerMaterial(float scale, float reflectiveness=0);
     PtrColor sample(PtrRay ray, PtrVector position, PtrVector normal);
 };
 
@@ -28,7 +28,7 @@ class PhongMaterial: public Material{
     PtrColor m_diffuse, m_specular;
     int m_shininess;
 public:
-    PhongMaterial(PtrColor diffuse, PtrColor specular, int shininess, double reflectiveness=0);
+    PhongMaterial(PtrColor diffuse, PtrColor specular, int shininess, float reflectiveness=0);
     PtrColor sample(PtrRay ray, PtrVector position, PtrVector normal);
     inline PtrColor getDiffuse(){ return m_diffuse; };
     inline PtrColor getSpecular(){ return m_specular; };
