+++ include/ray.hpp
@@ -8,7 +8,7 @@ class Ray{
 	PtrVector m_direction;
 public:
 	Ray(PtrVector origin, PtrVector direction);
-	PtrVector getPoint(double t);
+	PtrVector getPoint(float t);
 	inline PtrVector getOrigin(){ return m_origin; };
 	inline PtrVector getDirection(){ return m_direction; };
 };
