+++ src/perspectivecamera.cpp
@@ -3,7 +3,7 @@
 #include <vector.hpp>
 #include <ray.hpp>
 
-PerspectiveCamera::PerspectiveCamera(PtrVector eye, PtrVector front, PtrVector up, double fov)
+PerspectiveCamera::PerspectiveCamera(PtrVector eye, PtrVector front, PtrVector up, float fov)
     : m_eye(eye)
     , m_front(front)
     , m_refUp(up)
@@ -14,12 +14,12 @@ PerspectiveCamera::PerspectiveCamera(PtrVector eye, PtrVector front, PtrVector u
 void PerspectiveCamera::init(){
 	m_right = std::make_shared<Vector>(m_front->cross(*m_refUp));
 	m_up = std::make_shared<Vector>(m_right->cross(*m_front));
-	m_fovScale = tan(m_fov * 0.5 * PI / 180.0) * 2;
+	m_fovScale = tan(m_fov * 0.5f * PI / 180.0f) * 2.0f;
 }
 
-PtrRay PerspectiveCamera::generateRay(double x, double y){
-	Vector&& r = (*m_right) * ( (x - 0.5) * m_fovScale );
-	Vector&& u = (*m_up) * ( (y - 0.5) * m_fovScale );
+PtrRay PerspectiveCamera::generateRay(float x, float y){
+	Vector&& r = (*m_right) * ( (x - 0.5f) * m_fovScale );
+	Vector&& u = (*m_up) * ( (y - 0.5f) * m_fovScale );
 	Vector&& v = (*m_front+ r + u).normalize();
 	PtrRay ray = std::make_shared<Ray>(Ray(m_eye, std::make_shared<Vector>(v)));
 	return ray;
