+++ src/plane.cpp
@@ -4,7 +4,7 @@
 #include "ray.hpp"
 #include "intersect_result.hpp"
 
-Plane::Plane(PtrVector normal, double distance):
+Plane::Plane(PtrVector normal, float distance):
     m_normal(normal),
     m_distance(distance){
 
@@ -23,10 +23,10 @@ void Plane::init(){
     m_position = std::make_shared<Vector>(*m_normal * m_distance);
 }
 PtrIntersectResult Plane::intersect(PtrRay ray){
-    double dotA = ray->getDirection()->dot(*m_normal);
+    float dotA = ray->getDirection()->dot(*m_normal);
     if (dotA >= 0)
         return IntersectResult::NoHit;
-    double dotB = m_normal->dot(*(ray->getOrigin()) - *m_position);
-    double distance = -dotB / dotA;
+    float dotB = m_normal->dot(*(ray->getOrigin()) - *m_position);
+    float distance = -dotB / dotA;
     return std::make_shared<IntersectResult>(IntersectResult(shared_from_this(), distance, ray->getPoint(distance), m_normal));
 }
