+++ src/ray.cpp
@@ -7,7 +7,7 @@ Ray::Ray(PtrVector origin, PtrVector direction)
 {
 }
 
-PtrVector Ray::getPoint(double t)
+PtrVector Ray::getPoint(float t)
 {
     PtrVector result = std::make_shared<Vector>((*m_origin) + (*m_direction) * t);
     return result;
